Fix settings entry being injected multiple times
Not sure if this was broken in the last version as well.
Now the GUI for settings is built in 'Load()' and only updated in 'GetPanel()'. Simply switching the reference for the patch fixes the bug.

Added another patch to update the toggle when settings are opened. And moved the string key into a constant to prevent mess up.